- Summary: Fix Bugzilla bug 169030-Migrate to eclipse 3.3 SDK
- Bugzilla Bug (s) Resolved: 169030

- Description:
Fixed the BIRT 2.2.0 compiling errors on eclipse SDK 3.3 (GEF3.3, EMF2.3.0), JDK 1.5

- Tests Description: Manual test and code review

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

- Files Edited:
@changelist

- Files Added:
@addlist

- Files Deleted:
@deletelist

+++ UI/org.eclipse.birt.report.designer.ui/src/org/eclipse/birt/report/designer/internal/ui/editors/schematic/figures/LabelFigure.java
@@ -184,7 +184,8 @@ public class LabelFigure extends ReportElementFigure
 	{
 		if ( ft != null && ft.getFontData( ).length > 0 )
 		{
-			return ft.getFontData( )[0].height;
+			// return ft.getFontData( )[0].height;
+			return (int) ft.getFontData( )[0].height;
 		}
 
 		return 0;
